+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicAuthService.java
@@ -54,44 +54,4 @@ public class BasicAuthService implements AuthService {
                 currentUserPicture != null ? currentUserPicture.getBinaryData() : null
         );
     }
-
-    @Override
-    public void logOutUser(User user){
-        UserStatus userStatus = new UserStatus(user, false); // 새로 생성한 것을 저장 해야 하는데
-        userStatusRepository.createUserStatus(userStatus);
-    }
-
-    @Override
-    public UserStatus isOnline(User user){
-
-        UserStatus lastStatus = userStatusRepository.loadUserStatuses().stream()
-                .filter(userStatus -> userStatus.getUserId().equals(user.getId()))
-                .reduce((first, second) -> second)
-                .orElse(null); // 또는 예외 처리
-
-        assert lastStatus != null;
-        return lastStatus;
-
-        // true는 접속 시간
-        // faisl는
-    }
-
-    @Override
-    public List<UserLoggedDataDTO> getLoggedUserStatuses(){
-        /*
-        UserStatus
-        사용자 별 마지막으로 확인된 접속 시간을 표현하는 도메인 모델입니다. 사용자의 온라인 상태를 확인하기 위해 활용합니다.
-        마지막 접속 시간을 기준으로 현재 로그인한 유저로 판단할 수 있는 메소드를 정의하세요.
-        마지막 접속 시간이 현재 시간으로부터 5분 이내이면 현재 접속 중인 유저로 간주합니다.*/
-
-        List<UserLoggedDataDTO> userStatuses = new ArrayList<>();
-        List<User> usersFromFile = userRepository.loadUsers();
-
-        for (User user : usersFromFile) {
-            UserStatus userStatusFromCurrentUser = isOnline(user);
-            userStatuses.add(new UserLoggedDataDTO(user.getId(), user.getIdForLogin(), user.getUserName(), userStatusFromCurrentUser.getLoggedIn()));
-        }
-
-        return userStatuses;
-    }
 }
