+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -16,6 +16,7 @@ import org.springframework.stereotype.Service;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 
 @RequiredArgsConstructor // NOTE 생성자 lombok에서 생성 final이 붙은 필드를 매개변수로 받아 생성자에서 주입
@@ -25,7 +26,6 @@ public class BasicUserService implements UserService {
     private final ChannelRepository channelRepository;
     private final BinaryContentsRepository binaryContentsRepository;
     private final UserStatusRepository userStatusRepository;
-    private final AuthService authService;
 
     @Override
     public User createUser(UserCreateDTO userCreateDto) {
@@ -209,13 +209,49 @@ public class BasicUserService implements UserService {
                 .forEach(user -> System.out.printf("유저 이름: '%s', 유저 ID: '%s'\n", user.getUserName(), user.getId()));
     }
 
-    @Override
-    public List<UserLoggedDataDTO> findAll(){
-        return authService.getLoggedUserStatuses();
-    }
 
     @Override
     public UserLoggedDataDTO findUserById(String userIdForLogin){
         return findAll().stream().filter(u -> u.getUserId().equals(userIdForLogin)).findFirst().orElse(null);
+    }
+
+    @Override
+    public void logOutUser(User user){
+        UserStatus userStatus = new UserStatus(user, false); // 새로 생성한 것을 저장 해야 하는데
+        userStatusRepository.createUserStatus(userStatus);
+    }
+
+    @Override
+    public UserStatus isOnline(User user){
+
+        UserStatus lastStatus = userStatusRepository.loadUserStatuses().stream()
+                .filter(userStatus -> userStatus.getUserId().equals(user.getId()))
+                .reduce((first, second) -> second)
+                .orElse(null); // 또는 예외 처리
+
+        assert lastStatus != null;
+        return lastStatus;
+
+        // true는 접속 시간
+        // faisl는
+    }
+
+    @Override
+    public List<UserLoggedDataDTO> findAll(){
+        /*
+        UserStatus
+        사용자 별 마지막으로 확인된 접속 시간을 표현하는 도메인 모델입니다. 사용자의 온라인 상태를 확인하기 위해 활용합니다.
+        마지막 접속 시간을 기준으로 현재 로그인한 유저로 판단할 수 있는 메소드를 정의하세요.
+        마지막 접속 시간이 현재 시간으로부터 5분 이내이면 현재 접속 중인 유저로 간주합니다.*/
+
+        List<UserLoggedDataDTO> userStatuses = new ArrayList<>();
+        List<User> usersFromFile = userRepository.loadUsers();
+
+        for (User user : usersFromFile) {
+            UserStatus userStatusFromCurrentUser = isOnline(user);
+            userStatuses.add(new UserLoggedDataDTO(user.getId(), user.getIdForLogin(), user.getUserName(), userStatusFromCurrentUser.getLoggedIn()));
+        }
+
+        return userStatuses;
     }
 }
